Revert forced vsync
Compositor already uses "Mailbox" Vsync with 5 buffers, which is similar to Enhanced Sync from AMD, which is triple buffered.

It increases memory & CPU usage slightly compared to traditional double-buffered or triple-buffered Vsync, but it ensures better framepacing.

It is also lower in latency, as it dynamically discards frames which are not needed to display on screen.