Disable System76-Scheduler CFS tweaks
Linux 6.6 kernel is officially released & it will use EEVDF CPU scheduler by default instead of CFS, so this CFS modification is not needed.

Fedora 39 intended to ship 6.5 Linux kernel at 1st, but as things stand, it will probably ship Linux 6.6 kernel out-of-the-box, as Fedora release is delayed for 7th November.