Add support for untracking initramfs changes
This should solve the error when you're rebasing from 1 image to another.

When a file is tracked, but is not physically present, it will error out as "commit 123456789aetc not found"

As an example, when you switch from Surface image to the main one, it will error out because surface-modprobe file is not available.