Remove "fake-news" blocking hosts file
I didn't check on what it is, I thought it was just for very obvious examples of everyday fake-news & rumors, without political news attached to it.

But now that I looked into it better, it remains clear that the list is very-biased to the Western left-side of politics, which I'm not a fan of.

In fact, I'm not a fan of anyone politics.

If I find the non-political equivalent list, I will include it in the future.