extensions: Switch from my COPR RPMs to gnome-extensions module
It's much easier to maintain, where only change is to bump extension version number when new Fedora/Gnome releases.
Thanks to Gnome extensions website for supplying end-release instead of source, which makes it possible to have a working template installer script for almost all extensions.

Meanwhile,
With COPR I have to fork the extension,  install & configure Pull GitHub app, configure WebHook for automatic RPM build, fetch & merge the latest commits & go to slow COPR to check the status, which is a burden.

Same thing is if I want to build RPM based on release & not on git commits.

Besides that, I also have to check & test if RPM spec file is working, if source directory structure changed enough to break my RPM spec file (fixable by directly using make or meson when supplied, but I could not get it to work), etc.

OpenWeather Refined is the only extension I will maintain in RPM, as I applied as a maintainer for it. So that's the extension I won't remove here.