ISO: Disable flatpak integration
It doesn't work well.

It installs "Anaconda" flatpak ref incorrectly instead of using `flathub-system` that I specified.

Even with the working workaround of deleting the remote & adding it again, it's not the correct approach of flatpaks installation, so I will just switch to new ISO action & that's it.

If ISO flatpak installation improves in the future, I will switch to it.