Add "Unblank lock screen" extension
Testing it in a PR, because extension page says that it's automatically enabled when extension is installed & that it needs to be manually disabled.

I want it disabled by default, so I'm testing it here.

That note is probably only valid if extension is installed through DBUS (on booted system), while file-based method of gnome-extensions shouldn't modify anything regarding extension state.

But better to test it before landing.