Modify & add some more memory management tweaks
I was on good track with tweaks, except with `vm.swappiness`, which is hard to adjust per-device.

Adjusted minimum-free-zram (`vm.min_free_kbytes`) to account for CPU cores also, instead of generically using 1% of RAM as a value.

Added `vm.vfs_cache_pressure=66` instead of 100 tweak to improve responsiveness.

Thanks to @ahydronous & maxperfwiz for some of those advices:

https://gist.github.com/ahydronous/7ceaa00df96ef99131600edd4c2c73f2

https://gitlab.com/cscs/maxperfwiz