README: Remove `soar`
While `soar` itself is maintained, the main `pkgforge` repo is not, as the main developer went away, who was the only one who knew how to manage those complex automated workflows, so packages from there are all outdated.

Meanwhile, `AM` portable apps manager and it's "repo" is maintained, and even if it looks complex, it's much less complex compared to `pkgforge`. Also if it goes a bit unmaintained, packages won't be outdated if the download link doesn't change, while `pkgforge` builds packages directly, which stopped working.